refactor summary_file.py
